requirements.txt
openai
langchain_community
streamlit
python-dotenv
pypdf2
pandas
logging
-e .